refactor : apis.client 제거
