+++ apis/groups/queries.ts
@@ -1,4 +1,4 @@
-import { getGroups } from './apis.client';
+import { getGroup, getGroups } from './apis';
 import { Keys } from './keys';
 import { useQuery } from '@tanstack/react-query';
 
@@ -7,5 +7,5 @@ export const useGetGroups = (pageNum: number) => {
 };
 
 export const useGetGroup = (groupId: number) => {
-  return useQuery(Keys.getGroup(groupId), () => getGroups(groupId));
+  return useQuery(Keys.getGroup(groupId), () => getGroup(groupId));
 };
